Clean up more unused files
