--- src/lib/CalendarAbsolutePositionWrapper.svelte
@@ -1,26 +1,0 @@
-<div
-  style="
-    position: absolute; 
-    top: {calculatedTopPositionFromParent ? calculatedTopPositionFromParent : offsetFromTop}px;
-    left: {offsetFromLeft}px;
-  "
->
-  <slot>
-
-  </slot>
-</div>
-
-<script>
-  // import { computeOffset } from '/src/helpers/everythingElse.js'
-
-  // export let task
-  // export let pixelsPerHour
-  // export let calendarStartTime
-  export let offsetFromLeft
-  export let offsetFromTop
-
-  export let calculatedTopPositionFromParent
-
-  $: offsetFromTop = computeOffset(task, pixelsPerHour, calendarStartTime)
-  console.log('offsetFromTop =', offsetFromTop)
-</script>
